Crosswords

Grid utils

Grid utils

Enigmistics.create_grid — Function

create_grid(rows::Int, cols::Int; type="blank", probability=1.0, from=ALPHABET)

Create a grid of given number of rows and cols.

The argument type can either be "blank" (all empty cells) or "random" (grid randomly filled with density proportional to the given probability), while from indicates the set of characters to use when filling the grid randomly (default is ALPHABET, which contains only letters, otherwise there is EXT_ALPHABET which also contains black and empty cells).

Examples

julia> create_grid(4,4,type="blank")
4×4 Matrix{Char}:
 ' '  ' '  ' '  ' '
 ' '  ' '  ' '  ' '
 ' '  ' '  ' '  ' '
 ' '  ' '  ' '  ' '

julia> create_grid(4,4,type="random")
4×4 Matrix{Char}:
 'G'  'V'  'A'  'Y'
 'X'  'U'  'N'  'B'
 'X'  'Z'  'P'  'B'
 'J'  'E'  'Z'  'U'

julia> create_grid(4,4,type="random",probability=0.7, from=EXT_ALPHABET)
4×4 Matrix{Char}:
 ' '  'H'  'D'  ' '
 ' '  ' '  'H'  ' '
 'U'  'E'  ' '  'P'
 '■'  'B'  'C'  'A'

Enigmistics.show_grid — Function

show_grid(grid::Matrix{Char}; empty_placeholder = "⋅", style="single")

Show the grid in the console, with optional placeholder for empty cells and style of borders (either "single" or "double").

Examples

julia> g = create_grid(9,11,type="random", from=EXT_ALPHABET);

julia> show_grid(g,style="single")
     1  2  3  4  5  6  7  8  9 10 11 
   ┌─────────────────────────────────┐
1  │ Z  P  E  X  T  U  O  B  E  N  V │
2  │ V  A  R  R  D  N  Q  F  Z  I  T │
3  │ Y  U  R  ⋅  F  ⋅  S  Z  K  C  D │
4  │ D  N  Y  U  L  L  ■  S  G  Q  C │
5  │ D  D  W  V  Q  S  L  M  O  P  J │
6  │ ■  K  O  G  I  ■  D  Z  Q  A  ⋅ │
7  │ U  U  C  X  G  S  Y  C  F  W  E │
8  │ Q  D  V  W  G  O  O  B  Y  P  S │
9  │ L  H  Y  B  Q  Q  S  B  M  ■  S │
   └─────────────────────────────────┘

julia> show_grid(g,style="double", empty_placeholder = "_")
     1  2  3  4  5  6  7  8  9 10 11 
   ╔═════════════════════════════════╗
1  ║ Z  P  E  X  T  U  O  B  E  N  V ║
2  ║ V  A  R  R  D  N  Q  F  Z  I  T ║
3  ║ Y  U  R  _  F  _  S  Z  K  C  D ║
4  ║ D  N  Y  U  L  L  ■  S  G  Q  C ║
5  ║ D  D  W  V  Q  S  L  M  O  P  J ║
6  ║ ■  K  O  G  I  ■  D  Z  Q  A  _ ║
7  ║ U  U  C  X  G  S  Y  C  F  W  E ║
8  ║ Q  D  V  W  G  O  O  B  Y  P  S ║
9  ║ L  H  Y  B  Q  Q  S  B  M  ■  S ║
   ╚═════════════════════════════════╝

Enigmistics.insert_row_above — Function

insert_row_above(grid::Matrix{Char})

Take a grid and insert an empty row above it, returning the new grid.

Examples

julia> g = create_grid(3,3,type="random")
3×3 Matrix{Char}:
 'Z'  'D'  'S'
 'W'  'N'  'Q'
 'U'  'V'  'S'

julia> insert_row_above(g)
4×3 Matrix{Char}:
 ' '  ' '  ' '
 'Z'  'D'  'S'
 'W'  'N'  'Q'
 'U'  'V'  'S'

Enigmistics.insert_row_below — Function

insert_row_below(grid::Matrix{Char})

Take a grid and insert an empty row below it, returning the new grid.

Examples

julia> g = create_grid(3,3,type="random")
3×3 Matrix{Char}:
 'Z'  'D'  'S'
 'W'  'N'  'Q'
 'U'  'V'  'S'

julia> insert_row_below(g)
4×3 Matrix{Char}:
 'Z'  'D'  'S'
 'W'  'N'  'Q'
 'U'  'V'  'S'
 ' '  ' '  ' '

Enigmistics.insert_col_right — Function

insert_col_right(grid::Matrix{Char})

Take a grid and insert an empty column on the right of it, returning the new grid.

Examples

julia> g = create_grid(3,3,type="random")
3×3 Matrix{Char}:
 'Z'  'D'  'S'
 'W'  'N'  'Q'
 'U'  'V'  'S'

julia> insert_col_right(g)
3×4 Matrix{Char}:
 'Z'  'D'  'S'  ' '
 'W'  'N'  'Q'  ' '
 'U'  'V'  'S'  ' '

Enigmistics.insert_col_left — Function

insert_col_left(grid::Matrix{Char})

Take a grid and insert an empty column on the left of it, returning the new grid.

Examples

julia> g = create_grid(3,3,type="random")
3×3 Matrix{Char}:
 'Z'  'D'  'S'
 'W'  'N'  'Q'
 'U'  'V'  'S'

julia> insert_col_left(g)
3×4 Matrix{Char}:
 ' '  'Z'  'D'  'S'
 ' '  'W'  'N'  'Q'
 ' '  'U'  'V'  'S'

Enigmistics.insert — Function

insert(grid::Matrix{Char}, what::Symbol)

Take a grid and insert an empty row or column above/below or left/right of it, based on the symbol given by what (:row_above, :row_below, :col_left, :col_right), returning the new grid.

Enigmistics.shrink — Function

shrink(grid::Matrix{Char})

Take a grid and remove all-empty rows and columns from its borders. I.e., fits its contents to the smallest possible bounding box.

Examples

julia> using Random; Random.seed!(83)
julia> g = create_grid(7,7,type="random",probability=0.2)
7×7 Matrix{Char}:
 ' '  ' '  ' '  ' '  ' '  'C'  ' '
 ' '  ' '  ' '  ' '  ' '  ' '  'M'
 ' '  ' '  ' '  ' '  ' '  ' '  ' '
 ' '  ' '  'V'  ' '  ' '  ' '  ' '
 ' '  ' '  ' '  ' '  'A'  ' '  ' '
 ' '  ' '  ' '  'K'  ' '  ' '  ' '
 ' '  ' '  ' '  'Z'  ' '  ' '  ' '

julia> shrink(g)
7×5 Matrix{Char}:
 ' '  ' '  ' '  'C'  ' '
 ' '  ' '  ' '  ' '  'M'
 ' '  ' '  ' '  ' '  ' '
 'V'  ' '  ' '  ' '  ' '
 ' '  ' '  'A'  ' '  ' '
 ' '  'K'  ' '  ' '  ' '
 ' '  'Z'  ' '  ' '  ' '